# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Servanda Office API
  version: 1.0.0
  description: |
    Multi-Tenant LegalTech Platform API for contract creation and management.

    **Core Flow:** Template selection -> Guided Q&A Interview -> Conflict-free Contract -> DOCX/ODT Export.

    **Architecture Decisions:**
    - ADR-001: Shared DB with PostgreSQL Row-Level Security for tenant isolation
    - ADR-002: Immutable versioning for Clauses/Templates; ContractInstance pins versions
    - ADR-003: Async export via pgboss job queue with S3 storage
    - ADR-004: Server-side ODT conversion via LibreOffice headless

    **Authentication:**
    - Production: Bearer JWT (OIDC/SAML)
    - Development: Header-based tenant context (x-tenant-id, x-user-id, x-user-role)

    **Pagination:** All list endpoints support `page` and `pageSize` query parameters.
    Default page size: 20, maximum: 100.

    **Status Workflow (Clause/Template Versions):**
    `draft` -> `review` -> `approved` -> `published` -> `deprecated`
    (review -> draft and approved -> draft are also valid transitions)
  contact:
    name: Team 01 - Product Architecture
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server (v1 canonical prefix)
  - url: http://localhost:3000/api
    description: Legacy route prefix (backward-compatible)
security:
  - BearerAuth: []
  - DevHeaders: []
...
paths:
...
  /identity/users/{id}:
    get:
      tags: [Identity]
      summary: Get a single user
      description: Returns a user by ID within the current tenant.
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Identity]
      summary: Update a user
      description: Updates user role and/or display name. Admin role required.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Identity]
      summary: Delete a user
      description: |
        Permanently deletes a user. Admin role required.
        Cannot delete yourself.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Cannot delete yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production authentication via OIDC/SAML JWT token.
        Include tenant claim in the JWT payload.
    DevHeaders:
      type: apiKey
      in: header
      name: x-tenant-id
      description: |
        Development-mode authentication via headers.
        Requires all three headers:
        - `x-tenant-id` (UUID) - Tenant identifier
        - `x-user-id` (UUID) - User identifier
        - `x-user-role` (admin|editor|user) - User role

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request body or parameters (Zod validation failed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient permissions (role check failed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # =========================================================================
  # Schemas
  # =========================================================================
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    User:
      type: object
      required: [id, email, displayName, role, status, mfaEnabled, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [admin, editor, user]
        status:
          type: string
          enum: [invited, active, inactive]
        mfaEnabled:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, editor, user]
        displayName:
          type: string
          minLength: 1
          maxLength: 255
...